+++ packages/components/CHANGELOG.md
@@ -8,11 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.9.1-rc.0] - 2022-03-09
+
 #### Added
 - Styles for slotted `<button>` in `Text`
 
 #### Changed
 - `Modal` heading and aria validation happens only when open
+
+#### Fixed
+- React: bundling format of partials
 
 ### [2.9.0] - 2022-02-28